requirements.txt

torch
torchaudio
nemo_toolkit[asr]
hydra-core
omegaconf
sentencepiece
librosa
soundfile
numpy
jiwer
tqdm
pyyaml
requests
pandas